Move some arcane code into its own file to improve readability
